requirements.txt
aiogram==3.4.1
pydantic==2.5.3
pydantic-core==2.14.6
gunicorn==23.0.0
fastapi
python-dotenv
psycopg2-binary
asyncpg
openai
# json

